Загрузка прошлого кода в папку урока и корректировка кода исполняемых файлов в целях добавления логирования
